Define and export the HabularTable type
instead of creating an ad hoc object per table instance. Methods
of the HabularTable API are now part of the HabularTable prototype.
There is now _no_ state 'leaked' during the module's initialization
(previously a `table` variable was created to be later reused). The
exported habular object will now expose the `Table` (`HabularTable`)
constructor as well as the `table` convenience method which replaces
the previous `asTable`.